+++ crates/github_scbot_api/src/gif.rs
@@ -9,6 +9,7 @@ use serde::Deserialize;
 use crate::Result;
 
 const GIF_API_URL: &str = "https://g.tenor.com/v1";
+const MAX_GIF_SIZE_BYTES: usize = 2_000_000;
 
 #[allow(non_camel_case_types)]
 #[derive(Deserialize, PartialEq, Eq, Hash, Clone, Copy)]
@@ -31,6 +32,7 @@ enum GifFormat {
 #[derive(Deserialize)]
 struct MediaObject {
     url: String,
+    size: Option<usize>,
 }
 
 #[derive(Deserialize)]
@@ -43,18 +45,26 @@ struct RandomResponse {
     results: Vec<GifObject>,
 }
 
+const GIF_KEYS: &[GifFormat] = &[
+    GifFormat::Gif,
+    GifFormat::MediumGif,
+    GifFormat::TinyGif,
+    GifFormat::NanoGif,
+];
+
 /// Get random GIF for query.
-pub async fn random_gif_for_query(config: &Config, search: &str) -> Result<String> {
+pub async fn random_gif_for_query(config: &Config, search: &str) -> Result<Option<String>> {
     let client = reqwest::Client::new();
     let mut response: RandomResponse = client
         .get(&format!("{}/random", GIF_API_URL))
         .query(&[
             ("q", search),
             ("key", &config.tenor_api_key),
-            ("limit", "20"),
+            ("limit", "3"),
+            ("locale", "en_US"),
             ("contentfilter", "low"),
-            ("media_filter", "minimal"),
-            ("ar_range", "wide"),
+            ("media_filter", "basic"),
+            ("ar_range", "all"),
         ])
         .send()
         .await?
@@ -62,7 +72,7 @@ pub async fn random_gif_for_query(config: &Config, search: &str) -> Result<Strin
         .await?;
 
     if response.results.is_empty() {
-        Ok(String::new())
+        Ok(None)
     } else {
         let mut url = String::new();
 
@@ -73,13 +83,19 @@ pub async fn random_gif_for_query(config: &Config, search: &str) -> Result<Strin
         // Get first media found
         for result in &response.results {
             for media in &result.media {
-                if media.contains_key(&GifFormat::TinyGif) {
-                    url = media[&GifFormat::TinyGif].url.clone();
-                    break;
+                for key in GIF_KEYS {
+                    if media.contains_key(key) {
+                        if let Some(size) = media[key].size {
+                            if size < MAX_GIF_SIZE_BYTES {
+                                url = media[key].url.clone();
+                                break;
+                            }
+                        }
+                    }
                 }
             }
         }
 
-        Ok(url)
+        Ok(Some(url))
     }
 }