+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a `debug test-sentry` command to test Sentry connection
 - Manual interaction mode (using admin-enable command) to use the bot on specific pull requests only
+- Optional history tracking
 
 ### Fixed